+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Fixed
 
+- News posts now send emails via the job system to avoid timeouts
 - Avatar imports now work if the avatars aren't in the root of the ZIP file
 
 ## [4.2.0] - 2021-01-04